Update seed.py and recreate models files
Recreated user.py, profile.py and recipe.py
Changed from random value to fixed value during object creation.
Changed from writing results to a file to printing to the terminal.
