+++ server/models/profile.py
@@ -1,0 +1,35 @@
+from sqlalchemy import (
+    Column, String,
+    Integer, ForeignKey,
+    Text, Boolean
+    )
+from sqlalchemy.orm import relationship, validates
+from models.base_model import Base, BaseModelMixin
+
+class Profile(Base, BaseModelMixin):
+
+    __tablename__ = 'profile'
+
+    id = Column(Integer, primary_key=True)
+    name = Column(String, nullable=False)
+    is_chef = Column(Boolean, nullable=False)
+    about_me = Column(Text)
+    profile_image = Column(Text)
+    favourite_recipe = Column(Text)
+    favourite_cuisine = Column(String, nullable = False)
+    location = Column(Text, nullable = False)
+    user_id = Column(Integer, ForeignKey('user.id'))
+
+ 
+    #relationship Profile and Recipe
+    recipes = relationship("Recipe", backref="profile", cascade="all, delete-orphan") 
+
+    @validates('is_chef')
+    def validate_is_chef(self, key, is_chef):
+        if not isinstance(is_chef, bool):
+            raise AssertionError(f"Expected field 'is_chef' to be a boolean but got {type(is_chef)}")
+        return is_chef
+
+    def __repr__(self):
+        return f"<Profile #{self.id}: {self.name}>"
+
